backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic>=2.10.0
pydantic-settings>=2.7.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# AI & Agents
pydantic-ai
anthropic>=0.40.0
openai>=1.54.3

# Data Processing
pandas==2.2.0
numpy==1.26.3

# HTTP Clients
httpx>=0.27.2
aiohttp==3.9.1

# Utilities
python-dotenv==1.0.0
python-multipart>=0.0.10
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.1.0

# Background Tasks
celery==5.3.6
redis==5.0.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Development
black==23.12.1
ruff==0.1.11
mypy==1.8.0
griffe==2.0.0
python-jose[cryptography]
passlib[bcrypt]
python-multipart
